pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agaton-hackathon-base"
version = "0.1.0"
description = "Minimal multi-LLM agent base for hackathon"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-dotenv>=1.0.1",
  "openai>=1.59.0",
  "google-generativeai>=0.8.3",
  "anthropic>=0.42.0",
  "supabase>=2.11.0",
]

[project.optional-dependencies]
dev = ["ruff>=0.8.0"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = ["src", "README.md", "pyproject.toml"]
